:truck: Move depth buffer extraction to new class
