+++ src/Display/PointsForwardWarped.hpp
@@ -20,9 +20,8 @@ namespace csp::volumerendering {
 /// The number of used points is equal to the number of pixels in the used depth image.
 class PointsForwardWarped : public DisplayNode {
  public:
-  PointsForwardWarped(VolumeShape shape, std::shared_ptr<cs::core::Settings> settings,
-      std::string anchor, std::shared_ptr<cs::core::SolarSystem> solarSystem,
-      std::shared_ptr<cs::core::TimeControl> timeControl);
+  PointsForwardWarped(
+      VolumeShape shape, std::shared_ptr<cs::core::Settings> settings, std::string anchor);
 
   PointsForwardWarped(PointsForwardWarped const& other) = delete;
   PointsForwardWarped(PointsForwardWarped&& other)      = default;
