+++ src/Plugin.cpp
@@ -201,13 +201,14 @@ void Plugin::onLoad() {
     throw std::runtime_error("Failed to initialize CelestialObjects.");
   }
 
-  auto existence = anchor->second.mExistence;
-  mDisplayNodes[DisplayMode::eMesh] =
-      std::make_shared<Billboard>(mPluginSettings.mData.mShape.get(), mAllSettings,
-          mPluginSettings.mTransform.mAnchor.get(), mSolarSystem, mTimeControl);
-  mDisplayNodes[DisplayMode::ePoints] =
-      std::make_shared<PointsForwardWarped>(mPluginSettings.mData.mShape.get(), mAllSettings,
-          mPluginSettings.mTransform.mAnchor.get(), mSolarSystem, mTimeControl);
+  auto existence                    = anchor->second.mExistence;
+  mDisplayNodes[DisplayMode::eMesh] = std::make_shared<Billboard>(
+      mPluginSettings.mData.mShape.get(), mAllSettings, mPluginSettings.mTransform.mAnchor.get());
+  mDisplayNodes[DisplayMode::ePoints] = std::make_shared<PointsForwardWarped>(
+      mPluginSettings.mData.mShape.get(), mAllSettings, mPluginSettings.mTransform.mAnchor.get());
+
+  mDepthExtractor = std::make_unique<DepthExtractor>(
+      mDisplayNodes[DisplayMode::eMesh], mSolarSystem, mTimeControl);
 
   for (auto const& node : mDisplayNodes) {
     node.second->setAnchorPosition(mPluginSettings.mTransform.mPosition.get());
@@ -619,7 +620,7 @@ bool Plugin::tryRequestFrame() {
     dir           = dir * glm::inverse(mRenderingFrame.mCameraTransform);
     mRenderer->setSunDirection(dir);
     mFutureFrameData   = mRenderer->getFrame(mRenderingFrame.mCameraTransform,
-        std::move(mActiveDisplay->getDepthBuffer(mPluginSettings.mRendering.mResolution.get())));
+        std::move(mDepthExtractor->getDepthBuffer(mPluginSettings.mRendering.mResolution.get())));
     mLastFrameInterval = 0;
     mParametersDirty   = false;
     mFrameInvalid      = false;
@@ -751,7 +752,7 @@ std::vector<ScalarFilter> csp::volumerendering::Plugin::parseScalarFilters(
   std::vector<ScalarFilter> filters;
   for (auto const& [axis, value] : j.items()) {
     auto const& scalar = std::find_if(scalars.begin(), scalars.end(),
-        [&axis = axis](Scalar const& s) { return s.mName == axis; });
+        [& axis = axis](Scalar const& s) { return s.mName == axis; });
     if (scalar != scalars.end()) {
       ScalarFilter filter;
       filter.mAttrIndex = (int)std::distance(scalars.begin(), scalar);
